Rename hook files and add jsdoc comments
