+++ apps/client-react/src/views/project-edit-route.tsx
@@ -3,7 +3,7 @@ import { Link, useNavigate, useParams } from "react-router-dom";
 
 import { ymdToday } from "@projectsbuild/library/utils";
 import type { Project, ProjectForm, ProjectStatus } from "@projectsbuild/shared/types";
-import { useRerender } from "~/hooks/useRerender";
+import { useRerender } from "~/hooks/use-rerender";
 import { transformProject } from "./project-create-route";
 import { getProjectById } from "./project-route";
 import { useProjectsContext } from "./projects-route";
